♻️ BX2JXQ task: split task plan helpers
