+++ .agentplane/tasks/202604251626-BX2JXQ/README.md
@@ -1,10 +1,10 @@
 ---
 id: "202604251626-BX2JXQ"
 title: "Refactor task plan and finish spec surfaces"
-status: "TODO"
+status: "DOING"
 priority: "med"
 owner: "CODER"
-revision: 1
+revision: 6
 origin:
   system: "manual"
 depends_on:
@@ -16,19 +16,30 @@ verify:
   - "bun run docs:cli:check && bun run typecheck && bun run lint:core"
   - "bun run test:project -- agentplane packages/agentplane/src/cli/run-cli.core.lifecycle.plan.test.ts packages/agentplane/src/commands/task/finish.validation.unit.test.ts packages/agentplane/src/cli/run-cli.core.help-contract.test.ts"
 plan_approval:
-  state: "pending"
-  updated_at: null
-  updated_by: null
+  state: "approved"
+  updated_at: "2026-04-25T16:49:48.265Z"
+  updated_by: "ORCHESTRATOR"
   note: null
 verification:
   state: "pending"
   updated_at: null
   updated_by: null
   note: null
-comments: []
-events: []
+commit: null
+comments:
+  -
+    author: "CODER"
+    body: "Start: Refactor the task plan or finish command surface using one low-risk extraction, preserving command flags, generated help, and existing task behavior."
+events:
+  -
+    type: "status"
+    at: "2026-04-25T16:49:57.568Z"
+    author: "CODER"
+    from: "TODO"
+    to: "DOING"
+    note: "Start: Refactor the task plan or finish command surface using one low-risk extraction, preserving command flags, generated help, and existing task behavior."
 doc_version: 3
-doc_updated_at: "2026-04-25T16:26:36.613Z"
+doc_updated_at: "2026-04-25T16:54:11.084Z"
 doc_updated_by: "CODER"
 description: "Reduce task plan and finish spec hotspots by extracting spec option groups and render/validation helpers without changing command flags or generated help."
 sections:
@@ -40,9 +51,10 @@ sections:
     - In scope: Reduce task plan and finish spec hotspots by extracting spec option groups and render/validation helpers without changing command flags or generated help.
     - Out of scope: unrelated refactors not required for "Refactor task plan and finish spec surfaces".
   Plan: |-
-    1. Implement the change for "Refactor task plan and finish spec surfaces".
-    2. Run required checks and capture verification evidence.
-    3. Finalize task findings and finish with traceable commit metadata.
+    1. Inspect commands/task/plan.ts and commands/finish.spec.ts for the smallest low-risk extraction boundary.
+    2. Prefer one coherent extraction if both files together would widen scope; record any deferred half in Findings.
+    3. Preserve command flags, generated help, and task plan behavior.
+    4. Run focused plan/finish/help tests plus docs:cli check when command spec changes, then typecheck/lint/hotspot gates.
   Verify Steps: |-
     1. Run `bun run test:project -- agentplane packages/agentplane/src/cli/run-cli.core.lifecycle.plan.test.ts packages/agentplane/src/commands/task/finish.validation.unit.test.ts packages/agentplane/src/cli/run-cli.core.help-contract.test.ts`. Expected: it succeeds and confirms the requested outcome for this task.
     2. Run `bun run docs:cli:check && bun run typecheck && bun run lint:core`. Expected: it succeeds and confirms the requested outcome for this task.
@@ -54,7 +66,14 @@ sections:
   Rollback Plan: |-
     - Revert task-related commit(s).
     - Re-run required checks to confirm rollback safety.
-  Findings: ""
+  Findings: |-
+    - Observation: Extracted task plan backend and document validation helpers into plan-shared.ts; plan.ts is now below the runtime hotspot warning threshold while preserving command entrypoints.
+      Impact: Reduced the task plan command surface without changing generated help, flags, or plan approval behavior.
+      Resolution: Deferred finish.spec.ts extraction to a separate atom because its size is mostly declarative option and raw validation surface.
+    
+    - Observation: Full format:check remains blocked by unrelated untracked REFACTORING_PLAN_v3.md in the repository root.
+      Impact: This task can verify touched files with targeted Prettier, but cannot honestly claim the whole-repo format gate in the current dirty checkout.
+      Resolution: Left the untracked file untouched and used targeted Prettier checks for all files modified by this atom.
 id_source: "generated"
 ---
 ## Summary
@@ -70,9 +89,10 @@ Reduce task plan and finish spec hotspots by extracting spec option groups and r
 
 ## Plan
 
-1. Implement the change for "Refactor task plan and finish spec surfaces".
-2. Run required checks and capture verification evidence.
-3. Finalize task findings and finish with traceable commit metadata.
+1. Inspect commands/task/plan.ts and commands/finish.spec.ts for the smallest low-risk extraction boundary.
+2. Prefer one coherent extraction if both files together would widen scope; record any deferred half in Findings.
+3. Preserve command flags, generated help, and task plan behavior.
+4. Run focused plan/finish/help tests plus docs:cli check when command spec changes, then typecheck/lint/hotspot gates.
 
 ## Verify Steps
 
@@ -92,3 +112,11 @@ Reduce task plan and finish spec hotspots by extracting spec option groups and r
 - Re-run required checks to confirm rollback safety.
 
 ## Findings
+
+- Observation: Extracted task plan backend and document validation helpers into plan-shared.ts; plan.ts is now below the runtime hotspot warning threshold while preserving command entrypoints.
+  Impact: Reduced the task plan command surface without changing generated help, flags, or plan approval behavior.
+  Resolution: Deferred finish.spec.ts extraction to a separate atom because its size is mostly declarative option and raw validation surface.
+
+- Observation: Full format:check remains blocked by unrelated untracked REFACTORING_PLAN_v3.md in the repository root.
+  Impact: This task can verify touched files with targeted Prettier, but cannot honestly claim the whole-repo format gate in the current dirty checkout.
+  Resolution: Left the untracked file untouched and used targeted Prettier checks for all files modified by this atom.
